cli: extract start pkg
